+++ sonaranalyzer-dotnet/tests/SonarAnalyzer.UnitTest/TestCases/MethodShouldBeNamedAccordingToSynchronicity.MVC.Core.cs
